requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langsmith>=0.1.38
openai>=1.30.0
python-dotenv>=1.0.1
yfinance>=0.2.38
pandas>=2.2.2
requests>=2.31.0
sentence-transformers>=2.7.0
pgvector>=0.2.5
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.30
unstructured>=0.12.6
tiktoken>=0.7.0
streamlit>=1.31.0

# For PDF parsing (used in Agent 2 RAG)
pypdf>=4.2.0

# For News/Article scraping (used in Agent 1 tools)
newspaper3k>=0.2.8

# For DuckDuckGo tool (Agent 1 web search)
duckduckgo-search>=5.2.2
